[refactor] 기존 fcm 토큰 등록 api 의 request body 유효성 검증 강화 (#308)
- PlatformType 의 정적 팩토리 메서드를 활용하여 지원하는 FE 플랫폼인지 확인하도록 코드 수정
