+++ src/main/java/konkuk/thip/notification/adapter/in/web/request/FcmTokenRegisterRequest.java
@@ -3,8 +3,8 @@ package konkuk.thip.notification.adapter.in.web.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import konkuk.thip.notification.domain.value.PlatformType;
 import konkuk.thip.notification.application.port.in.dto.FcmTokenRegisterCommand;
+import konkuk.thip.common.util.PlatformType;
 
 @Schema(description = "FCM 토큰 등록 요청 DTO")
 public record FcmTokenRegisterRequest(
@@ -18,13 +18,13 @@ public record FcmTokenRegisterRequest(
 
         @NotNull
         @Schema(description = "플랫폼 타입 (ANDROID 또는 WEB)", example = "ANDROID")
-        PlatformType platformType
+        String platformType
 ) {
     public FcmTokenRegisterCommand toCommand(Long userId) {
         return FcmTokenRegisterCommand.builder()
                 .deviceId(this.deviceId)
                 .fcmToken(this.fcmToken)
-                .platformType(this.platformType)
+                .platformType(PlatformType.from(this.platformType))
                 .userId(userId)
                 .build();
     }
