+++ src/main/java/konkuk/thip/notification/application/port/in/dto/FcmTokenRegisterCommand.java
@@ -1,6 +1,6 @@
 package konkuk.thip.notification.application.port.in.dto;
 
-import konkuk.thip.notification.domain.value.PlatformType;
+import konkuk.thip.common.util.PlatformType;
 import lombok.Builder;
 
 @Builder
